requirements.txt
langgraph==0.4.5
pandas==2.2.3
openpyxl==3.1.5
chromadb==1.0.9
openai==1.79.0
uvicorn==0.34.2
fastapi==0.115.9
loguru==0.7.3
python-docx==1.1.2
langchain==0.3.25
upstash-redis
python-dotenv==1.0.1
pydantic-settings==2.3.4
jinja2==3.1.4
